Cleanedup runner code
