+++ caribou/src/caribou/execution/report_generation.py
@@ -1,0 +1,99 @@
+"""
+Agent report generation and session statistics tracking.
+
+This module handles:
+- Generating handoff reports between agents
+- Managing agent report memory for multi-agent workflows
+- Writing session reports to disk
+"""
+from __future__ import annotations
+
+import json
+from datetime import datetime
+from pathlib import Path
+from typing import Dict, List, Optional
+
+from rich.console import Console
+
+
+class AgentReportMemory:
+    """
+    Minimal memory that keeps global system prompts, the current agent prompt,
+    and a collection of handoff reports generated by prior agents.
+    """
+    def __init__(self, global_messages: List[Dict[str, str]], agent_prompt: str) -> None:
+        self._global_messages = list(global_messages)
+        self._agent_prompt = agent_prompt
+        self._agent_reports: List[Dict[str, str]] = []
+
+    def update_agent_prompt(self, new_prompt: str) -> None:
+        self._agent_prompt = new_prompt
+
+    def add_report(self, agent_name: str, report: str) -> None:
+        content = f"Agent report from {agent_name}:\n{report}"
+        self._agent_reports.append({"role": "system", "content": content})
+
+    def build_context(self, working_history: List[Dict[str, str]]) -> List[Dict[str, str]]:
+        context: List[Dict[str, str]] = []
+        context.extend(self._global_messages)
+        if self._agent_prompt:
+            context.append({"role": "system", "content": self._agent_prompt})
+        context.extend(self._agent_reports)
+        context.extend(working_history)
+        return context
+
+
+def _write_session_report(
+    console: Console,
+    *,
+    output_dir: Path,
+    stats: Dict[str, object],
+) -> Optional[Path]:
+    """Persist session statistics to disk."""
+    report_dir = output_dir / "reports"
+    report_dir.mkdir(parents=True, exist_ok=True)
+    report_path = report_dir / f"session_report_{datetime.utcnow().strftime('%Y%m%d-%H%M%S')}.json"
+    try:
+        report_path.write_text(json.dumps(stats, indent=2))
+        console.print(f"[bold green]✓ Session report saved to:[/bold green] {report_path}")
+        return report_path
+    except Exception as exc:
+        console.print(f"[yellow]Warning: Failed to write session report: {exc}[/yellow]")
+        return None
+
+
+def _generate_agent_report(
+    console: Console,
+    *,
+    llm_client: object,
+    model_name: str,
+    agent_name: str,
+    history_slice: List[Dict[str, str]],
+) -> Optional[str]:
+    """
+    Ask the LLM to produce a concise handoff report for the completed agent.
+    Emphasis is on data generated, file changes, and key outcomes.
+    """
+    if not history_slice:
+        return "No history available for this agent; nothing to report."
+
+    report_prompt = (
+        "You are producing a handoff report for the previous agent in a multi-agent workflow. "
+        "Summarize what the agent did, focusing on data that was generated or modified, code that was executed or changed, "
+        "outputs produced (with paths if mentioned), and any warnings or blockers. "
+        "Be concise but specific so the next agent can pick up where this one left off."
+    )
+
+    try:
+        resp = llm_client.chat.completions.create(
+            model=model_name,
+            messages=[
+                {"role": "system", "content": report_prompt},
+                {"role": "user", "content": json.dumps(history_slice)},
+            ],
+            temperature=0.3,
+        )
+        return resp.choices[0].message.content
+    except Exception as exc:
+        console.print(f"[yellow]Warning: Failed to generate agent report: {exc}[/yellow]")
+        return None
